chore(expo): Preserve relative imports via neverBundle in tsdown
