+++ .changeset/expo-preserve-relative-imports.md
@@ -1,0 +1,5 @@
+---
+'@clerk/expo': patch
+---
+
+Preserve Expo's relative import specifiers in tsdown builds so Metro platform-specific module resolution and root exports work correctly.
