--- .changeset/expo-restore-tsup-build.md
@@ -1,5 +1,0 @@
----
-'@clerk/expo': patch
----
-
-Restore the Expo package build to tsup to preserve Metro platform-specific module resolution and root exports.
